Refactor: feature 단위로 분리 (MM-174)
